+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Incorrect behavior of `Scene.setBackground`
 - Removed deprecated method `Story.registerScene`
+- Condition actions are not executed correctly
 
 ## [0.8.1]