+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Incorrect behavior of `Scene.setBackground`
 - Removed deprecated method `Story.registerScene`
 - Condition actions are not executed correctly
+- Null action execution can cause infinite loop
 
 ## [0.8.1]